fix(wrapper,v17,phase5a): add state-layer validation to init_lp_vault_registry/init_lp_redemption; migrate lp_vault tests to v17 API
REAL BUG: init_lp_vault_registry and init_lp_redemption silently accepted
invalid inputs (zero market_group, zero lp_mint, zero registry, zero redeemer,
fee_share_bps > 10_000, bad version, paused > 1). The handler validates at the
program boundary but the state layer must be self-protecting (defense-in-depth).
Added explicit field guards to both init functions — exposed by the state_tests
suite which was correctly testing these invariants.

v17 API migrations (test adjustments only, behavior preserved):
- lp_vault_state_tests: TAG_CREATE_LP_VAULT 65→74, TAG_CLOSE_LP_VAULT 71→80
  (v17 renumber, matrix rows 21-27, avoids toly tag 65-73 collisions)
- lp_vault_deposit: TopUpBackingBucket domain u8 cast removed (domain field is
  now u16 in v17 per auth overhaul)
- lp_vault_redeem: RequestRedeemLpShares lp_amount → shares field rename;
  same domain u16 cast removal for TopUpBackingBucket/WithdrawBackingBucket

+++ tests/v16_fork_lp_vault_deposit.rs
@@ -483,7 +483,7 @@ fn lp_deposit_backing_state_matches_top_up() {
         &mut env_a.svm,
         env_a.program_id,
         &env_a.payer,
-        ProgInstruction::TopUpBackingBucket { domain: DOMAIN as u8, amount: DEPOSIT, expiry_slot: LP_VAULT_BACKING_EXPIRY_SLOT },
+        ProgInstruction::TopUpBackingBucket { domain: DOMAIN, amount: DEPOSIT, expiry_slot: LP_VAULT_BACKING_EXPIRY_SLOT },
         vec![
             AccountMeta::new(admin_a.pubkey(), true),
             AccountMeta::new(env_a.market, false),
